Prevent writing all marked imgs onto file (lossage) (#534)
This fix issues a warning message and does nothing in that case.

Running C-x c (copy) or C-x m (move) will copy/move all the marked files to a
single location. If that location is not a directory, then creating that file
and overwriting it several times results in complete loss of data in the case of
moving, and unhelpful behaviour in the case of "copying". Only the last marked
file remains in the file.